model_deployment/requirements.txt
Flask==2.0.1
Werkzeug==2.0.1
SQLAlchemy==1.4.23
psycopg2-binary==2.9.1